requirements.txt
celery
python-memcached
pyyaml
http://www.freedesktop.org/~lanius/pyxdg-0.18.tar.gz
